Major re-write/cleanup of the questions classes
